Renamed application context & moved Aside component
