+++ sites/partners/pages/application/edit.tsx
@@ -13,6 +13,7 @@ import { ApplicationStatus } from "@bloom-housing/backend-core/types"
 import Layout from "../../layouts/application"
 import PaperApplicationForm from "../../src/applications/PaperApplicationForm/PaperApplicationForm"
 import { useSingleApplicationData } from "../../lib/hooks"
+import { ApplicationContext } from "../../src/applications/ApplicationContext"
 
 const NewApplication = () => {
   const metaDescription = ""
@@ -26,35 +27,41 @@ const NewApplication = () => {
   if (!application) return false
 
   return (
-    <Layout>
-      <Head>
-        <title>{t("nav.siteTitle")}</title>
-      </Head>
-      <MetaTags title={t("nav.siteTitle")} image={metaImage} description={metaDescription} />
-      <PageHeader>
-        {t("applications.editApplication")}{" "}
-        <p className="text-base">{`${application.applicant?.firstName} ${application.applicant?.middleName} ${application.applicant?.lastName}`}</p>
-      </PageHeader>
+    <ApplicationContext.Provider value={application}>
+      <Layout>
+        <Head>
+          <title>{t("nav.siteTitle")}</title>
+        </Head>
+        <MetaTags title={t("nav.siteTitle")} image={metaImage} description={metaDescription} />
+        <PageHeader>
+          {t("applications.editApplication")}{" "}
+          <p className="text-base">{`${application.applicant?.firstName} ${application.applicant?.middleName} ${application.applicant?.lastName}`}</p>
+        </PageHeader>
 
-      <StatusBar
-        backButton={
-          <Button inlineIcon="left" icon="arrow-back" onClick={() => router.back()}>
-            {t("t.back")}
-          </Button>
-        }
-        tagStyle={
-          application?.status == ApplicationStatus.submitted
-            ? AppearanceStyleType.success
-            : AppearanceStyleType.primary
-        }
-        tagLabel={
-          application?.status
-            ? t(`application.details.applicationStatus.${application.status}`)
-            : t(`application.details.applicationStatus.submitted`)
-        }
-      />
-      <PaperApplicationForm listingId={application.listing.id} application={application} editMode />
-    </Layout>
+        <StatusBar
+          backButton={
+            <Button inlineIcon="left" icon="arrow-back" onClick={() => router.back()}>
+              {t("t.back")}
+            </Button>
+          }
+          tagStyle={
+            application?.status == ApplicationStatus.submitted
+              ? AppearanceStyleType.success
+              : AppearanceStyleType.primary
+          }
+          tagLabel={
+            application?.status
+              ? t(`application.details.applicationStatus.${application.status}`)
+              : t(`application.details.applicationStatus.submitted`)
+          }
+        />
+        <PaperApplicationForm
+          listingId={application.listing.id}
+          application={application}
+          editMode
+        />
+      </Layout>
+    </ApplicationContext.Provider>
   )
 }
