+++ sites/partners/src/applications/ApplicationContext.ts
@@ -1,6 +1,6 @@
 import React from "react"
 import { Application } from "@bloom-housing/backend-core/types"
 
-export const DetailsApplicationContext = React.createContext<Application | null>(null)
+export const ApplicationContext = React.createContext<Application | null>(null)
 
-DetailsApplicationContext.displayName = "DetailsApplicationContext"
+ApplicationContext.displayName = "ApplicationContext"
