+++ sites/partners/src/applications/PaperApplicationDetails/sections/DetailsHouseholdMembers.tsx
@@ -1,6 +1,6 @@
 import React, { useContext, useMemo } from "react"
 import { t, GridSection, MinimalTable, Button } from "@bloom-housing/ui-components"
-import { DetailsApplicationContext } from "../../ApplicationContext"
+import { ApplicationContext } from "../../ApplicationContext"
 import { MembersDrawer } from "../DetailsMemberDrawer"
 import { YesNoAnswer } from "../../PaperApplicationForm/FormTypes"
 
@@ -9,7 +9,7 @@ type DetailsHouseholdMembersProps = {
 }
 
 const DetailsHouseholdMembers = ({ setMembersDrawer }: DetailsHouseholdMembersProps) => {
-  const application = useContext(DetailsApplicationContext)
+  const application = useContext(ApplicationContext)
 
   const householdMembersHeaders = {
     name: t("t.name"),
