+++ sites/partners/src/applications/PaperApplicationDetails/sections/DetailsPreferences.tsx
@@ -1,9 +1,9 @@
 import React, { useContext } from "react"
 import { t, GridSection, ViewItem, GridCell } from "@bloom-housing/ui-components"
-import { DetailsApplicationContext } from "../../ApplicationContext"
+import { ApplicationContext } from "../../ApplicationContext"
 
 const DetailsPreferences = () => {
-  const application = useContext(DetailsApplicationContext)
+  const application = useContext(ApplicationContext)
 
   return (
     <GridSection className="bg-primary-lighter" title={t("application.details.preferences")} inset>
