+++ sites/partners/src/applications/PaperApplicationDetails/sections/DetailsPrimaryApplicant.tsx
@@ -1,11 +1,11 @@
 import React, { useContext } from "react"
 import { t, GridSection, ViewItem, GridCell } from "@bloom-housing/ui-components"
-import { DetailsApplicationContext } from "../../ApplicationContext"
+import { ApplicationContext } from "../../ApplicationContext"
 import { DetailsAddressColumns, AddressColsType } from "../DetailsAddressColumns"
 import { YesNoAnswer } from "../../PaperApplicationForm/FormTypes"
 
 const DetailsPrimaryApplicant = () => {
-  const application = useContext(DetailsApplicationContext)
+  const application = useContext(ApplicationContext)
 
   return (
     <GridSection
